# proto/openapi.yaml
openapi: 3.1.0
info:
  title: SeaSnail OpenAPI
  version: 0.1.0
  description: |
    SeaSnail 本地音频转译服务的 HTTP/REST API。本地 UI 与第三方应用统一经此 API 接入（OpenAPI 为唯一集成面）。

    ## 鉴权
    Scoped Bearer Token，请求头 `Authorization: Bearer <token>`。token 独立于用户密码，可吊销、可多签、**绑 account_id**。
    可用 scope：
      - `sessions:read`    读取 / 搜索会话、取原始音频、导出
      - `sessions:write`  创建会话、编辑转译文本（**仅本地 UI token 可得，不对第三方签发**）
      - `sessions:delete` 删除会话（可按需签发给第三方）
      - `tokens:manage`    管理 token（**仅本地 UI token 可得**）

    **多账户模型**：MVP 支持多账户，每账户独立密码 + 独立 master DEK，账户间密钥完全隔离。日常免密只对**活跃账户**成立（其 DEK 缓存于 app 私有 keychain，`last_active_account_id` 记于非敏感 prefs 供重开取回）；切换账户需重新认证（`POST /accounts/{id}/unlock`）。

    - 首个账户：`POST /auth/setup`（免鉴权、仅 loopback、仅"未初始化"态可用）建首个账户并签发该账户 **root token**（`is_root=true` + `account_id`，持有全部 scope），root token secret 存 app 私有 keychain，本地 UI 据此调用。
    - 后续账户：`POST /accounts`（需任一已解锁账户的 root token）——无全局 root，靠"已解锁即可建新账户"（已确认）。
    - 切换/解锁：`POST /accounts/{id}/unlock`（免鉴权 + password）→ 解 wrap 该账户 DEK 入 keychain 作活跃 → 签发/恢复该账户 root token。
    - 第三方 token 经 `POST /tokens` 签发，**绑 account_id**（仅可访问该账户、且仅该账户已解锁时），**非 root 调用方不可被授予 `sessions:write` / `tokens:manage`**（daemon 硬约束），故第三方仅可得 `sessions:read` / `sessions:delete`。
    - 改密走 `POST /auth/password`（需 root token + 当前密码）。

    本地优先：服务监听本机；bootstrap 不走云端，token 由本地 UI 解锁后在 Integrations 面板生成。

    ## 数据安全
    入库数据（原始音频 + 转译内容）经加密落库（per-account 独立密钥）；读取 / 导出时由服务解密后返回。
...
    导出为敏感操作，需在请求体中提供用户密码再确认（作用于活跃账户）。
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ss_live_
  responses:
...
    Unauthorized:
      description: 未携带或无效 token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 权限不足（缺 scope / 密码错误）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 当前资源状态与请求冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Gone:
      description: 一次性资源已消费、过期或不可再用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: insufficient_scope }
            message: { type: string }
          required: [code, message]
      required: [error]
...
    Token:
      type: object
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, description: 所属账户（token 绑定此 account_id，仅可访问该账户且仅该账户已解锁时） }
        name: { type: string, description: 用户命名 }
        prefix: { type: string, example: ss_live_abcd…, description: 仅展示前缀 }
        is_root: { type: boolean, description: 是否 root token（每账户自己的 root，is_root + account_id；仅 setup/unlock 签发，持有 write/manage scope） }
        scopes:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time, nullable: true }
      required: [id, account_id, name, scopes, created_at]
...
    TokenCreated:
      allOf:
        - $ref: '#/components/schemas/Token'
        - type: object
          properties:
            secret: { type: string, description: 完整 token 明文，仅创建时返回一次 }
          required: [secret]
    AccountCreate:
      type: object
      description: 新建账户（独立密码 + 独立 master DEK，与其它账户密钥完全隔离）
      properties:
        username: { type: string }
        password: { type: string, format: password }
      required: [username, password]
    AccountSummary:
      type: object
      properties:
        id: { type: string, description: 不透明、稳定的账户标识（改名不影响密钥） }
        username: { type: string }
        created_at: { type: string, format: date-time }
        is_active: { type: boolean, description: 是否当前活跃账户（DEK 在 keychain） }
      required: [id, username, created_at, is_active]
...
security:
  - bearerAuth: []
paths:
...
  /auth/setup:
    post:
      summary: 建首个账户并签发该账户 root token
      description: |
        免鉴权，仅 loopback，仅"未初始化"态可用（已有任何账户则返回 410）。
        建首个账户：接收用户名+密码 → KDF→KEK→随机 master_DEK[acct]→wrapped-DEK[acct] 落盘 + master_DEK[acct] 存 app 私有 keychain（作活跃）
        + 记 last_active_account_id → 签发该账户 root token（is_root=true + account_id，全 scope），明文 secret 仅返回一次。
        后续账户改走 POST /accounts（需任一已解锁账户的 root token）。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string, format: password }
              required: [username, password]
      responses:
        '201':
          description: 已初始化
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenCreated' }
        '410': { $ref: '#/components/responses/Gone' }
  /auth/password:
    post:
      summary: 修改密码（重 wrap master_DEK[acct]，DEK 与历史内容不变）
      description: |
        需 root token（tokens:manage 仅 root 持有）+ 当前密码（知识因子，双重确认）。作用于 caller 的 account_id 账户。
        校验当前密码（解开 wrapped-DEK[acct] 验证）→ 新密码 KDF 重 wrap master_DEK[acct] 落盘。master_DEK 与历史内容不变。
        不影响 token：token secret 独立随机、不派生自密码，改密后所有 token 仍有效；要撤销某 token 走 DELETE /tokens/{id}。
      x-required-scope: tokens:manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password }
              required: [current_password, new_password]
      responses:
        '200': { description: 已修改 }
        '403': { $ref: '#/components/responses/Forbidden' }
  /accounts:
    get:
      summary: 列出账户
      description: 需 root token。返回所有账户摘要（含 is_active 标记活跃账户）。
      x-required-scope: tokens:manage
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AccountSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: 追加账户（独立密码 + 独立 DEK，切换活跃）
      description: |
        需任一已解锁账户的 root token（无全局 root，靠"已解锁即可建新账户"——已确认）。
        建独立账户：独立密码 → 独立 KEK → 独立 master_DEK[acct] → 独立 wrapped-DEK[acct] 落盘
        → 切换活跃至新账户（DEK 入 keychain + 记 last_active_account_id）
        → 签发该新账户 root token，明文 secret 仅返回一次。
      x-required-scope: tokens:manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '201':
          description: 已创建并切为活跃
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenCreated' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /accounts/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    delete:
      summary: 删除账户（删该账户全部数据）
      description: |
        需 root token。删该账户的 DB + 文件树 + wrapped-DEK + keychain 项。
        活跃账户删除后回退到 setup/选账户态；活跃账户不可直接删（先切走），否则 409。
      x-required-scope: tokens:manage
      responses:
        '204': { description: 已删除 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /accounts/{id}/unlock:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      summary: 切换/解锁账户
      description: |
        免鉴权（密码作知识因子）+ 仅 loopback。输该账户密码 → KDF→KEK→解 wrapped-DEK[acct] 验证
        → master_DEK[acct] 入 keychain 作活跃（替换旧活跃项）→ 签发该账户 root token
        → 更新 last_active_account_id。旧活跃账户 DEK 移出 keychain（其 wrapped-DEK 留盘）。
        明文 root token secret 仅返回一次，并缓存 keychain 供日常免密。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string, format: password }
              required: [password]
      responses:
        '200':
          description: 已解锁并切为活跃
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenCreated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...